market/requirements.txt
celery==5.1.2
Django==3.2.4
django-rest-passwordreset==1.2.0
djangorestframework==3.12.4
drf-spectacular==0.17.2
model-bakery==1.3.2
psycopg2-binary==2.9.1
pytest==6.2.4
pytest-django==4.4.0
PyYAML==5.4.1
redis==3.5.3
ujson==4.0.2










